# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Colombo Relay ITS API",
        "description": "API for managing ITS passes and related information for Colombo Relay.",
        "contact": {
            "email": "admin@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://vms-api-main-branch-zuipth.laravel.cloud",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/hizbe-saifee-groups": {
            "get": {
                "tags": [
                    "Hizbe Saifee Groups"
                ],
                "summary": "Get list of Hizbe Saifee groups",
                "description": "Returns list of Hizbe Saifee groups. Admin access required.",
                "operationId": "getHizbeSaifeeGroupsList",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/HizbeSaifeeGroup"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden (Admin access required)"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Hizbe Saifee Groups"
                ],
                "summary": "Update an existing Hizbe Saifee group",
                "description": "Updates an existing Hizbe Saifee group. Admin access required.",
                "operationId": "updateHizbeSaifeeGroup",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateHizbeSaifeeGroupRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/HizbeSaifeeGroup"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden (Admin access required)"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Hizbe Saifee Groups"
                ],
                "summary": "Store a new Hizbe Saifee group",
                "description": "Stores a new Hizbe Saifee group. Admin access required.",
                "operationId": "storeHizbeSaifeeGroup",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreHizbeSaifeeGroupRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/HizbeSaifeeGroup"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden (Admin access required)"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/hizbe-saifee-groups/{id}": {
            "get": {
                "tags": [
                    "Hizbe Saifee Groups"
                ],
                "summary": "Get Hizbe Saifee group information",
                "description": "Returns Hizbe Saifee group data. Admin access required.",
                "operationId": "getHizbeSaifeeGroupById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Hizbe Saifee Group id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HizbeSaifeeGroup"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden (Admin access required)"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Hizbe Saifee Groups"
                ],
                "summary": "Delete an existing Hizbe Saifee group",
                "description": "Deletes an existing Hizbe Saifee group. Admin access required.",
                "operationId": "deleteHizbeSaifeeGroup",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Hizbe Saifee Group id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                },
                                "example": {
                                    "success": true,
                                    "message": "Hizbe Saifee Group deleted successfully."
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden (Admin access required)"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "StoreHizbeSaifeeGroupRequest": {
                "title": "Store Hizbe Saifee Group Request",
                "required": [
                    "name",
                    "capacity",
                    "group_no"
                ],
                "properties": {
                    "name": {
                        "description": "Name of the group",
                        "type": "string",
                        "example": "Al-Ameen Group"
                    },
                    "capacity": {
                        "description": "Capacity of the group",
                        "type": "integer",
                        "example": 50
                    },
                    "group_no": {
                        "description": "Unique group number",
                        "type": "integer",
                        "example": 101
                    },
                    "whatsapp_link": {
                        "description": "WhatsApp group link",
                        "type": "string",
                        "example": "https://chat.whatsapp.com/samplelink",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "UpdateHizbeSaifeeGroupRequest": {
                "title": "Update Hizbe Saifee Group Request",
                "required": [
                    "id"
                ],
                "properties": {
                    "id": {
                        "description": "ID of the group to update",
                        "type": "integer",
                        "format": "int64"
                    },
                    "name": {
                        "description": "Name of the group",
                        "type": "string",
                        "example": "Al-Ameen Group Updated"
                    },
                    "capacity": {
                        "description": "Capacity of the group",
                        "type": "integer",
                        "example": 55
                    },
                    "group_no": {
                        "description": "Unique group number",
                        "type": "integer",
                        "example": 102
                    },
                    "whatsapp_link": {
                        "description": "WhatsApp group link",
                        "type": "string",
                        "example": "https://chat.whatsapp.com/newlink",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "HizbeSaifeeGroup": {
                "title": "Hizbe Saifee Group",
                "description": "Hizbe Saifee Group model",
                "properties": {
                    "id": {
                        "description": "ID",
                        "type": "integer",
                        "format": "int64"
                    },
                    "name": {
                        "description": "Name of the group",
                        "type": "string"
                    },
                    "capacity": {
                        "description": "Capacity of the group",
                        "type": "integer"
                    },
                    "group_no": {
                        "description": "Unique group number",
                        "type": "integer"
                    },
                    "whatsapp_link": {
                        "description": "WhatsApp group link",
                        "type": "string",
                        "nullable": true
                    },
                    "created_at": {
                        "description": "Creation timestamp",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Last update timestamp",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter token in format (Bearer <token>)",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}